requirements.txt
Django==2.2.6
django-constrainedfilefield==3.2.0
django-guardian==2.1.0
Pillow==6.2.0
psycopg2==2.8.3
python-magic==0.4.15
pytz==2019.2
sqlparse==0.3.0
